Moving RPCSIDReferentConformantArray
