--- src/main/java/com/rapid7/client/dcerpc/objects/RPCSIDReferentConformantArray.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2017, Rapid7, Inc.
- *
- * License: BSD-3-clause
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *   Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- *  Redistributions in binary form must reproduce the above copyright
- * notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution.
- *
- *  Neither the name of the copyright holder nor the names of its contributors
- * may be used to endorse or promote products derived from this software
- * without specific prior written permission.
- *
- *
- */
-package com.rapid7.client.dcerpc.objects;
-
-import com.rapid7.client.dcerpc.io.ndr.arrays.RPCReferentConformantArray;
-
-public class RPCSIDReferentConformantArray extends RPCReferentConformantArray<RPCSID> {
-
-    public RPCSIDReferentConformantArray(RPCSID[] array) {
-        super(array);
-    }
-
-    @Override
-    protected RPCSID[] createArray(int length) {
-        return new RPCSID[length];
-    }
-
-    @Override
-    protected RPCSID createEntry() {
-        return new RPCSID();
-    }
-}
